Extracted IPP to HAL for morphology functions
