+++ hal/ipp/CMakeLists.txt
@@ -30,6 +30,7 @@ add_library(ipphal STATIC
     "${CMAKE_CURRENT_SOURCE_DIR}/src/hough_ipp.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/accum_ipp.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/corner_ipp.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/morph_ipp.cpp"
 )
 
 #TODO: HAVE_IPP_ICV and HAVE_IPP_IW added as private macro till OpenCV itself is
