+++ modules/imgproc/src/morph.dispatch.cpp
@@ -264,216 +264,7 @@ static bool halMorph(int op, int src_type, int dst_type,
     return success;
 }
 
-// ===== 2. IPP implementation
-#if 0 //defined HAVE_IPP
-#ifdef HAVE_IPP_IW
-static inline IwiMorphologyType ippiGetMorphologyType(int morphOp)
-{
-    return morphOp == MORPH_ERODE ? iwiMorphErode :
-        morphOp == MORPH_DILATE   ? iwiMorphDilate :
-        morphOp == MORPH_OPEN     ? iwiMorphOpen :
-        morphOp == MORPH_CLOSE    ? iwiMorphClose :
-        morphOp == MORPH_GRADIENT ? iwiMorphGradient :
-        morphOp == MORPH_TOPHAT   ? iwiMorphTophat :
-        morphOp == MORPH_BLACKHAT ? iwiMorphBlackhat : (IwiMorphologyType)-1;
-}
-#endif
-
-static bool ippMorph(int op, int src_type, int dst_type,
-              const uchar * src_data, size_t src_step,
-              uchar * dst_data, size_t dst_step,
-              int width, int height,
-              int roi_width, int roi_height, int roi_x, int roi_y,
-              int roi_width2, int roi_height2, int roi_x2, int roi_y2,
-              int kernel_type, uchar * kernel_data, size_t kernel_step,
-              int kernel_width, int kernel_height, int anchor_x, int anchor_y,
-              int borderType, const double borderValue[4], int iterations, bool isSubmatrix)
-{
-#ifdef HAVE_IPP_IW
-    CV_INSTRUMENT_REGION_IPP();
-
-#if IPP_VERSION_X100 < 201800
-    // Problem with SSE42 optimizations performance
-    if(cv::ipp::getIppTopFeatures() == ippCPUID_SSE42)
-        return false;
-
-    // Different mask flipping
-    if(op == MORPH_GRADIENT)
-        return false;
-
-    // Integer overflow bug
-    if(src_step >= IPP_MAX_32S ||
-       src_step*height >= IPP_MAX_32S)
-        return false;
-#endif
-
-#if IPP_VERSION_X100 < 201801
-    // Problem with AVX512 optimizations performance
-    if(cv::ipp::getIppTopFeatures()&ippCPUID_AVX512F)
-        return false;
-
-    // Multiple iterations on small mask is not effective in current integration
-    // Inplace imitation for 3x3 kernel is not efficient
-    // Advanced morphology for small mask introduces degradations
-    if((iterations > 1 || src_data == dst_data || (op != MORPH_ERODE && op != MORPH_DILATE)) && kernel_width*kernel_height < 25)
-        return false;
-
-    // Skip even mask sizes for advanced morphology since they can produce out of spec writes
-    if((op != MORPH_ERODE && op != MORPH_DILATE) && (!(kernel_width&1) || !(kernel_height&1)))
-        return false;
-#endif
-
-    IppAutoBuffer<Ipp8u>        kernelTempBuffer;
-    ::ipp::IwiBorderSize        iwBorderSize;
-    ::ipp::IwiBorderSize        iwBorderSize2;
-    ::ipp::IwiBorderType        iwBorderType;
-    ::ipp::IwiBorderType        iwBorderType2;
-    ::ipp::IwiImage             iwMask;
-    ::ipp::IwiImage             iwInter;
-    ::ipp::IwiSize              initSize(width, height);
-    ::ipp::IwiSize              kernelSize(kernel_width, kernel_height);
-    IppDataType                 type        = ippiGetDataType(CV_MAT_DEPTH(src_type));
-    int                         channels    = CV_MAT_CN(src_type);
-    IwiMorphologyType           morphType   = ippiGetMorphologyType(op);
-
-    CV_UNUSED(isSubmatrix);
-
-    if((int)morphType < 0)
-        return false;
-
-    if(iterations > 1 && morphType != iwiMorphErode && morphType != iwiMorphDilate)
-        return false;
-
-    if(src_type != dst_type)
-        return false;
-
-    if(!ippiCheckAnchor(anchor_x, anchor_y, kernel_width, kernel_height))
-        return false;
-
-    try
-    {
-        ::ipp::IwiImage iwSrc(initSize, type, channels, ::ipp::IwiBorderSize(roi_x, roi_y, roi_width-roi_x-width, roi_height-roi_y-height), (void*)src_data, src_step);
-        ::ipp::IwiImage iwDst(initSize, type, channels, ::ipp::IwiBorderSize(roi_x2, roi_y2, roi_width2-roi_x2-width, roi_height2-roi_y2-height), (void*)dst_data, dst_step);
-
-        iwBorderSize = ::ipp::iwiSizeToBorderSize(kernelSize);
-        iwBorderType = ippiGetBorder(iwSrc, borderType, iwBorderSize);
-        if(!iwBorderType)
-            return false;
-        if(iterations > 1)
-        {
-            // Check dst border for second and later iterations
-            iwBorderSize2 = ::ipp::iwiSizeToBorderSize(kernelSize);
-            iwBorderType2 = ippiGetBorder(iwDst, borderType, iwBorderSize2);
-            if(!iwBorderType2)
-                return false;
-        }
-
-        if(morphType != iwiMorphErode && morphType != iwiMorphDilate && morphType != iwiMorphGradient)
-        {
-            // For now complex morphology support only InMem around all sides. This will be improved later.
-            if((iwBorderType&ippBorderInMem) && (iwBorderType&ippBorderInMem) != ippBorderInMem)
-                return false;
-
-            if((iwBorderType&ippBorderInMem) == ippBorderInMem)
-            {
-                iwBorderType &= ~ippBorderInMem;
-                iwBorderType &=  ippBorderFirstStageInMem;
-            }
-        }
-
-        if(iwBorderType.StripFlags() == ippBorderConst)
-        {
-            if(Vec<double, 4>(borderValue) == morphologyDefaultBorderValue())
-                iwBorderType.SetType(ippBorderDefault);
-            else
-                iwBorderType.m_value = ::ipp::IwValueFloat(borderValue[0], borderValue[1], borderValue[2], borderValue[3]);
-        }
-
-        iwMask.Init(ippiSize(kernel_width, kernel_height), ippiGetDataType(CV_MAT_DEPTH(kernel_type)), CV_MAT_CN(kernel_type), 0, kernel_data, kernel_step);
-
-        ::ipp::IwiImage iwMaskLoc = iwMask;
-        if(morphType == iwiMorphDilate)
-        {
-            iwMaskLoc.Alloc(iwMask.m_size, iwMask.m_dataType, iwMask.m_channels);
-            ::ipp::iwiMirror(iwMask, iwMaskLoc, ippAxsBoth);
-            iwMask = iwMaskLoc;
-        }
-
-        if(iterations > 1)
-        {
-            // OpenCV uses in mem border from dst for two and more iterations, so we need to keep this border in intermediate image
-            iwInter.Alloc(initSize, type, channels, iwBorderSize2);
-
-            ::ipp::IwiImage *pSwap[2] = {&iwInter, &iwDst};
-            CV_INSTRUMENT_FUN_IPP(::ipp::iwiFilterMorphology, iwSrc, iwInter, morphType, iwMask, ::ipp::IwDefault(), iwBorderType);
-
-            // Copy border only
-            {
-                if(iwBorderSize2.top)
-                {
-                    ::ipp::IwiRoi   borderRoi(-iwBorderSize2.left, -iwBorderSize2.top, iwDst.m_size.width+iwBorderSize2.left+iwBorderSize2.right, iwBorderSize2.top);
-                    ::ipp::IwiImage iwInterRoi = iwInter.GetRoiImage(borderRoi);
-                    ::ipp::iwiCopy(iwDst.GetRoiImage(borderRoi), iwInterRoi);
-                }
-                if(iwBorderSize2.bottom)
-                {
-                    ::ipp::IwiRoi   borderRoi(-iwBorderSize2.left, iwDst.m_size.height, iwDst.m_size.width+iwBorderSize2.left+iwBorderSize2.right, iwBorderSize2.bottom);
-                    ::ipp::IwiImage iwInterRoi = iwInter.GetRoiImage(borderRoi);
-                    ::ipp::iwiCopy(iwDst.GetRoiImage(borderRoi), iwInterRoi);
-                }
-                if(iwBorderSize2.left)
-                {
-                    ::ipp::IwiRoi   borderRoi(-iwBorderSize2.left, 0, iwBorderSize2.left, iwDst.m_size.height);
-                    ::ipp::IwiImage iwInterRoi = iwInter.GetRoiImage(borderRoi);
-                    ::ipp::iwiCopy(iwDst.GetRoiImage(borderRoi), iwInterRoi);
-                }
-                if(iwBorderSize2.right)
-                {
-                    ::ipp::IwiRoi   borderRoi(iwDst.m_size.width, 0, iwBorderSize2.left, iwDst.m_size.height);
-                    ::ipp::IwiImage iwInterRoi = iwInter.GetRoiImage(borderRoi);
-                    ::ipp::iwiCopy(iwDst.GetRoiImage(borderRoi), iwInterRoi);
-                }
-            }
-
-            iwBorderType2.SetType(iwBorderType);
-            for(int i = 0; i < iterations-1; i++)
-                CV_INSTRUMENT_FUN_IPP(::ipp::iwiFilterMorphology, *pSwap[i&0x1], *pSwap[(i+1)&0x1], morphType, iwMask, ::ipp::IwDefault(), iwBorderType2);
-            if(iterations&0x1)
-                CV_INSTRUMENT_FUN_IPP(::ipp::iwiCopy, iwInter, iwDst);
-        }
-        else
-        {
-            if(src_data == dst_data)
-            {
-                iwInter.Alloc(initSize, type, channels);
-
-                CV_INSTRUMENT_FUN_IPP(::ipp::iwiFilterMorphology, iwSrc, iwInter, morphType, iwMask, ::ipp::IwDefault(), iwBorderType);
-                CV_INSTRUMENT_FUN_IPP(::ipp::iwiCopy, iwInter, iwDst);
-            }
-            else
-                CV_INSTRUMENT_FUN_IPP(::ipp::iwiFilterMorphology, iwSrc, iwDst, morphType, iwMask, ::ipp::IwDefault(), iwBorderType);
-        }
-    }
-    catch(const ::ipp::IwException &)
-    {
-        return false;
-    }
-
-    return true;
-#else
-    CV_UNUSED(op); CV_UNUSED(src_type); CV_UNUSED(dst_type); CV_UNUSED(src_data); CV_UNUSED(src_step); CV_UNUSED(dst_data);
-    CV_UNUSED(dst_step); CV_UNUSED(width); CV_UNUSED(height); CV_UNUSED(roi_width); CV_UNUSED(roi_height);
-    CV_UNUSED(roi_x); CV_UNUSED(roi_y); CV_UNUSED(roi_width2); CV_UNUSED(roi_height2); CV_UNUSED(roi_x2); CV_UNUSED(roi_y2);
-    CV_UNUSED(kernel_type); CV_UNUSED(kernel_data); CV_UNUSED(kernel_step); CV_UNUSED(kernel_width); CV_UNUSED(kernel_height);
-    CV_UNUSED(anchor_x); CV_UNUSED(anchor_y); CV_UNUSED(borderType); CV_UNUSED(borderValue); CV_UNUSED(iterations);
-    CV_UNUSED(isSubmatrix);
-    return false;
-#endif
-};
-
-#endif // HAVE_IPP
-
-// ===== 3. Fallback implementation
+// ===== 2. Fallback implementation
 
 static void ocvMorph(int op, int src_type, int dst_type,
                      uchar * src_data, size_t src_step,
@@ -538,13 +329,6 @@ void morph(int op, int src_type, int dst_type,
         if (res)
             return;
     }
-
-    /*CV_IPP_RUN_FAST(ippMorph(op, src_type, dst_type, src_data, src_step, dst_data, dst_step, width, height,
-                            roi_width, roi_height, roi_x, roi_y,
-                            roi_width2, roi_height2, roi_x2, roi_y2,
-                            kernel_type, kernel_data, kernel_step,
-                            kernel_width, kernel_height, anchor_x, anchor_y,
-                            borderType, borderValue, iterations, isSubmatrix));*/
 
     ocvMorph(op, src_type, dst_type, src_data, src_step, dst_data, dst_step, width, height,
              roi_width, roi_height, roi_x, roi_y,
@@ -1128,76 +912,6 @@ static bool ocl_morphologyEx(InputArray _src, OutputArray _dst, int op,
 
 #endif
 
-#define IPP_DISABLE_MORPH_ADV 1
-#if 0 //defined HAVE_IPP
-#if !IPP_DISABLE_MORPH_ADV
-static bool ipp_morphologyEx(int op, InputArray _src, OutputArray _dst,
-                     InputArray _kernel,
-                     Point anchor, int iterations,
-                     int borderType, const Scalar& borderValue)
-{
-#if defined HAVE_IPP_IW
-    Mat kernel = _kernel.getMat();
-    Size ksize = !kernel.empty() ? kernel.size() : Size(3,3);
-    anchor = normalizeAnchor(anchor, ksize);
-
-    if (iterations == 0 || kernel.rows*kernel.cols == 1)
-    {
-        _src.copyTo(_dst);
-        return true;
-    }
-
-    if (kernel.empty())
-    {
-        kernel = getStructuringElement(MORPH_RECT, Size(1+iterations*2,1+iterations*2));
-        anchor = Point(iterations, iterations);
-        iterations = 1;
-    }
-    else if( iterations > 1 && countNonZero(kernel) == kernel.rows*kernel.cols )
-    {
-        anchor = Point(anchor.x*iterations, anchor.y*iterations);
-        kernel = getStructuringElement(MORPH_RECT,
-                                       Size(ksize.width + (iterations-1)*(ksize.width-1),
-                                            ksize.height + (iterations-1)*(ksize.height-1)),
-                                       anchor);
-        iterations = 1;
-    }
-
-    Mat src = _src.getMat();
-    _dst.create( src.size(), src.type() );
-    Mat dst = _dst.getMat();
-
-    Point s_ofs;
-    Size s_wsz(src.cols, src.rows);
-    Point d_ofs;
-    Size d_wsz(dst.cols, dst.rows);
-    bool isolated = (borderType&BORDER_ISOLATED)?true:false;
-    borderType = (borderType&~BORDER_ISOLATED);
-
-    if(!isolated)
-    {
-        src.locateROI(s_wsz, s_ofs);
-        dst.locateROI(d_wsz, d_ofs);
-    }
-
-    return ippMorph(op, src.type(), dst.type(),
-               src.data, src.step,
-               dst.data, dst.step,
-               src.cols, src.rows,
-               s_wsz.width, s_wsz.height, s_ofs.x, s_ofs.y,
-               d_wsz.width, d_wsz.height, d_ofs.x, d_ofs.y,
-               kernel.type(), kernel.data, kernel.step, kernel.cols, kernel.rows, anchor.x, anchor.y,
-               borderType, borderValue.val, iterations,
-               (src.isSubmatrix() && !isolated));
-#else
-    CV_UNUSED(op); CV_UNUSED(_src); CV_UNUSED(_dst); CV_UNUSED(_kernel); CV_UNUSED(anchor);
-    CV_UNUSED(iterations); CV_UNUSED(borderType); CV_UNUSED(borderValue);
-    return false;
-#endif
-}
-#endif
-#endif
-
 void morphologyEx( InputArray _src, OutputArray _dst, int op,
                        InputArray _kernel, Point anchor, int iterations,
                        int borderType, const Scalar& borderValue )
@@ -1224,10 +938,6 @@ void morphologyEx( InputArray _src, OutputArray _dst, int op,
     Mat src = _src.getMat(), temp;
     _dst.create(src.size(), src.type());
     Mat dst = _dst.getMat();
-
-#if !IPP_DISABLE_MORPH_ADV
-    //CV_IPP_RUN_FAST(ipp_morphologyEx(op, src, dst, kernel, anchor, iterations, borderType, borderValue));
-#endif
 
     switch( op )
     {
